requirements.txt
pandas==1.1.4
dash==1.21.0
dash-bootstrap-components==0.13.0
plotly-express==0.4.1 
folium==0.12.1
numpy==1.19.3
regex==2021.10.23
gunicorn==20.1.0